search_portal/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
ecdsa==0.19.1
email-validator==2.1.0
environ==1.0
exceptiongroup==1.3.0
fastapi==0.104.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
inflection==0.5.1
motor==3.3.2
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.9.0
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.27.0
typing_extensions==4.14.1
uritemplate==4.2.0
uvicorn==0.24.0
whitenoise==6.9.0
